pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "vics-agent"
version = "0.1.0"
description = "Coding your day away by Vics — An autonomous coding agent"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Vics" },
]
keywords = ["agent", "ai", "coding", "autonomous", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
vics = "vics_agent.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
